Move theme-classic/src/utils code to new @docusaurus/theme-common package
