+++ packages/docusaurus-theme-classic/src/theme/LayoutProviders/index.tsx
@@ -8,7 +8,7 @@
 import React from 'react';
 import ThemeProvider from '@theme/ThemeProvider';
 import UserPreferencesProvider from '@theme/UserPreferencesProvider';
-import DocsPreferredVersionContextProvider from '../../utils/docsPreferredVersion/DocsPreferredVersionProvider';
+import {DocsPreferredVersionContextProvider} from '@docusaurus/theme-common';
 
 export default function LayoutProviders({children}) {
   return (
