+++ packages/docusaurus-theme-common/src/utils/docsUtils.ts
