+++ packages/docusaurus-theme-common/src/utils/pathUtils.ts
