+++ packages/docusaurus-theme-common/src/utils/searchUtils.ts
